Stop generation for the whole convo
