+++ front/lib/api/assistant/generation.ts
@@ -412,10 +412,14 @@ export async function* runGeneration(
   const _checkCancellation = async () => {
     try {
       const cancelled = await redis.get(
-        `assistant:generation:cancelled:${agentMessage.sId}`
+        `assistant:generation:cancelled:${conversation.sId}`
       );
-      if (cancelled) {
+      if (cancelled === "1") {
         shouldYieldCancel = true;
+        await redis.set(
+          `assistant:generation:cancelled:${conversation.sId}`,
+          0
+        );
       }
     } catch (error) {
       console.error("Error checking cancellation:", error);
