+++ front/pages/api/w/[wId]/assistant/conversations/[cId]/events.ts
@@ -1,14 +1,27 @@
+import { isLeft } from "fp-ts/lib/Either";
+import * as t from "io-ts";
+import * as reporter from "io-ts-reporters";
 import { NextApiRequest, NextApiResponse } from "next";
 
 import { getConversation } from "@app/lib/api/assistant/conversation";
-import { getConversationEvents } from "@app/lib/api/assistant/pubsub";
+import {
+  cancelConversationGenerationEvent,
+  getConversationEvents,
+} from "@app/lib/api/assistant/pubsub";
 import { Authenticator, getSession } from "@app/lib/auth";
 import { ReturnedAPIErrorType } from "@app/lib/error";
 import { apiError, withLogging } from "@app/logger/withlogging";
 
+export type PostMessageEventResponseBody = {
+  success: true;
+};
+const PostMessageEventBodySchema = t.type({
+  action: t.literal("cancel"),
+});
+
 async function handler(
   req: NextApiRequest,
-  res: NextApiResponse<ReturnedAPIErrorType>
+  res: NextApiResponse<PostMessageEventResponseBody | ReturnedAPIErrorType>
 ): Promise<void> {
   const session = await getSession(req, res);
   const auth = await Authenticator.fromSession(
@@ -104,12 +117,29 @@ async function handler(
       res.status(200).end();
       return;
 
+    case "POST":
+      const bodyValidation = PostMessageEventBodySchema.decode(req.body);
+      if (isLeft(bodyValidation)) {
+        const pathError = reporter.formatValidationErrors(bodyValidation.left);
+
+        return apiError(req, res, {
+          status_code: 400,
+          api_error: {
+            type: "invalid_request_error",
+            message: `Invalid request body: ${pathError}`,
+          },
+        });
+      }
+      await cancelConversationGenerationEvent(conversation.sId);
+      return res.status(200).json({ success: true });
+
     default:
       return apiError(req, res, {
         status_code: 405,
         api_error: {
           type: "method_not_supported_error",
-          message: "The method passed is not supported, GET is expected.",
+          message:
+            "The method passed is not supported, GET or POST is expected.",
         },
       });
   }
