+++ front/pages/api/w/[wId]/assistant/conversations/[cId]/messages/[mId]/events.ts
@@ -1,28 +1,15 @@
-import { isLeft } from "fp-ts/lib/Either";
-import * as t from "io-ts";
-import * as reporter from "io-ts-reporters";
 import { NextApiRequest, NextApiResponse } from "next";
 
 import { getConversation } from "@app/lib/api/assistant/conversation";
-import {
-  cancelMessageGenerationEvent,
-  getMessagesEvents,
-} from "@app/lib/api/assistant/pubsub";
+import { getMessagesEvents } from "@app/lib/api/assistant/pubsub";
 import { Authenticator, getSession } from "@app/lib/auth";
 import { ReturnedAPIErrorType } from "@app/lib/error";
 import { apiError, withLogging } from "@app/logger/withlogging";
 import { isAgentMessageType } from "@app/types/assistant/conversation";
 
-export type postMessageEventResponseBody = {
-  success: true;
-};
-const PostMessageEventBodySchema = t.type({
-  action: t.literal("cancel"),
-});
-
 async function handler(
   req: NextApiRequest,
-  res: NextApiResponse<postMessageEventResponseBody | ReturnedAPIErrorType>
+  res: NextApiResponse<ReturnedAPIErrorType>
 ): Promise<void> {
   const session = await getSession(req, res);
   const auth = await Authenticator.fromSession(
@@ -153,29 +140,12 @@ async function handler(
       res.status(200).end();
       return;
 
-    case "POST":
-      const bodyValidation = PostMessageEventBodySchema.decode(req.body);
-      if (isLeft(bodyValidation)) {
-        const pathError = reporter.formatValidationErrors(bodyValidation.left);
-
-        return apiError(req, res, {
-          status_code: 400,
-          api_error: {
-            type: "invalid_request_error",
-            message: `Invalid request body: ${pathError}`,
-          },
-        });
-      }
-      await cancelMessageGenerationEvent(conversation.sId, messageId);
-      return res.status(200).json({ success: true });
-
     default:
       return apiError(req, res, {
         status_code: 405,
         api_error: {
           type: "method_not_supported_error",
-          message:
-            "The method passed is not supported, GET or POST is expected.",
+          message: "The method passed is not supported, GET is expected.",
         },
       });
   }
